bugfix: replace AuthCheckService with the AuthCheck guard middleware to make tests pass
