+++ .env
@@ -1,4 +1,5 @@
 PORT=4000
+TEST_MODE=auth
 
 CRYPT_SALT=10
 JWT_SECRET_KEY=secret123123
