+++ package.json
@@ -34,6 +34,7 @@
     "@nestjs/core": "^10.4.7",
     "@nestjs/jwt": "^10.2.0",
     "@nestjs/mapped-types": "^2.0.6",
+    "@nestjs/passport": "^10.0.3",
     "@nestjs/platform-express": "^10.4.7",
     "@nestjs/swagger": "^8.0.7",
     "@prisma/client": "^5.22.0",
