+++ src/endpoints/auth/auth.controller.ts
@@ -6,6 +6,7 @@ import {
   ClassSerializerInterceptor,
   HttpCode,
 } from '@nestjs/common';
+
 import { AuthService } from './auth.service';
 import { CreateUserDto } from '../user/user.entity';
