+++ src/endpoints/auth/auth.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
+import { JwtModule } from '@nestjs/jwt';
+
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { UserService } from '../user/user.service';
-import { JwtModule } from '@nestjs/jwt';
 import { TokenExpiryTimeEnum } from '../../helpers/enums';
 import { PrismaService } from '../../prisma/prisma.service';
