+++ src/main.ts
@@ -1,9 +1,8 @@
 import { ValidationPipe } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
-import { AuthCheckService } from './tools/auth-check/auth-check.service';
-import { TestAuthCheckService } from './tools/auth-check/test-auth-check.service';
 import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
+
+import { AppModule } from './app.module';
 
 async function bootstrap() {
   BigInt.prototype['toJSON'] = function () {
@@ -19,10 +18,6 @@ async function bootstrap() {
   const document = SwaggerModule.createDocument(app, config);
   SwaggerModule.setup('doc', app, document);
 
-  await app
-    .useGlobalPipes(new ValidationPipe())
-    .use(new TestAuthCheckService().use)
-    .use(new AuthCheckService(app.get(AuthCheckService)).use)
-    .listen(process.env.PORT);
+  await app.useGlobalPipes(new ValidationPipe()).listen(process.env.PORT);
 }
 bootstrap().then();
