--- src/tools/auth-check/auth-check.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AuthCheckService } from './auth-check.service';
-import { JwtService } from '@nestjs/jwt';
-
-@Module({
-  providers: [AuthCheckService, JwtService],
-  exports: [AuthCheckService],
-})
-export class AuthCheckModule {}
